Langchain-crash-course/QA-app-pinecone-Gemini-pro/requirements (2).txt
streamlit
pinecone-client
langchain
pypdf
pandas
numpy
python-dotenv
langchain-pinecone
langchain-community
langchain-google-genai
nest-asyncio
